Properly support DATABASE_URL
Switched out the code for loading configuration for Rails's version of
loading database configuration.

Rails has implemented support for DATABASE_URL in rails/rails@13463 and
has dealt with the possible combinations of having DATABASE_URL set and
a YAML config file, see https://github.com/rails/rails/pull/13463#issuecomment-31480799.

There is a breaking change in the way the hash is loaded; Rails does not
convert the keys of the hash to symbols, instead they are left as
strings; any code reliant on the hash keys being symbols will break.